Handle setting inverse values that are non-hashable (#23)
By default, when we do inverse mapping on values, `list`, `dict`, and
`set` values are not hashable and throw an error.

This PR casts them first to a string, both in the getter and setter
methods.

I don't think we need to worry about collisions, since I think
effectively lists `[1,2,3] == [1,2,3]` even if they are different
instantiations. I can't think of a use case where you'd have identical
lists of different IDs that you'd want to be keyed differently.