Handle non-typed and mixed type inputs (#25)
The previous implementation was a bit too insistent upon inspecting
input arguments of specific types (strings, to be exact).

This refactors it to be less strict. Now, if annotated dependencies are
found, they are hydrated; if not, arguments are provided in order they
are passed.

Probably will need to revisit this further to support potential `kwargs`
to a run function.