Refactor graph execution (#111)
This refactors the way Diagraph approaches graph execution.

Instead of getting groups of nodes to operate on, and concurrently
executing all nodes in a group (and waiting for that node group to
complete), it now takes a list of pointers to the starting nodes of a
graph and steps those pointers through the graph, executing any nodes it
fines as soon as they are ready to execute.

This slightly deprecates the concept of "layers" in a diagraph, but
should speed up execution.